Correctly handle file and directory literals produced by expression tool in output.
