+++ cwltool/load_tool.py
@@ -105,6 +105,8 @@ def validate_document(document_loader, workflowobj, uri,
     fileuri = urlparse.urldefrag(uri)[0]
 
     if "cwlVersion" in workflowobj:
+        if not isinstance(workflowobj["cwlVersion"], basestring):
+            raise Exception("'cwlVersion' must be a string, got %s" % type(workflowobj["cwlVersion"]))
         workflowobj["cwlVersion"] = re.sub(
             r"^(?:cwl:|https://w3id.org/cwl/cwl#)", "",
             workflowobj["cwlVersion"])
