+++ cwltool/process.py
@@ -12,6 +12,7 @@ from collections import Iterable
 import errno
 import shutil
 import uuid
+import hashlib
 
 import abc
 import schema_salad.validate as validate
@@ -201,13 +202,12 @@ def relocateOutputs(outputObj, outdir, output_dirs, action):
 
     def moveIt(src, dst):
         for a in output_dirs:
-            if src.startswith(a):
-                if action == "move":
-                    _logger.debug("Moving %s to %s", src, dst)
-                    shutil.move(src, dst)
-                elif action == "copy":
-                    _logger.debug("Copying %s to %s", src, dst)
-                    shutil.copy(src, dst)
+            if action == "move" and src.startswith(a):
+                _logger.debug("Moving %s to %s", src, dst)
+                shutil.move(src, dst)
+                continue
+            _logger.debug("Copying %s to %s", src, dst)
+            shutil.copy(src, dst)
 
     outfiles = []  # type: List[Dict[Text, Any]]
     collectFilesAndDirs(outputObj, outfiles)
@@ -216,6 +216,10 @@ def relocateOutputs(outputObj, outdir, output_dirs, action):
 
     def _check_adjust(f):
         f["location"] = "file://" + pm.mapper(f["location"])[1]
+        if "contents" in f:
+            del f["contents"]
+        if f["class"] == "File":
+            compute_checksums(StdFsAccess(""), f)
         return f
 
     adjustFileObjs(outputObj, _check_adjust)
@@ -700,3 +704,16 @@ def scandeps(base, doc, reffields, urlfields, loadref):
         normalizeFilesDirs(r)
         r = mergedirs(r)
     return r
+
+def compute_checksums(fs_access, fileobj):
+    if "checksum" not in fileobj:
+        checksum = hashlib.sha1()
+        with fs_access.open(fileobj["location"], "rb") as f:
+            contents = f.read(1024*1024)
+            while contents != "":
+                checksum.update(contents)
+                contents = f.read(1024*1024)
+            f.seek(0, 2)
+            filesize = f.tell()
+        fileobj["checksum"] = "sha1$%s" % checksum.hexdigest()
+        fileobj["size"] = filesize
